# perf: Don't Sprintf for ordered lists